+++ NEWS.md
@@ -1,6 +1,7 @@
 # ggplotify 0.0.4
 
-+ `as.grob` method for `eulergram` (defined in eulerr package) object (2019-05-09, Thu)
++ `as.grob` method for `Heatmap` object (defined in ComplexHeatmap package) (2019-08-06, Tue)
++ `as.grob` method for `eulergram` object (defined in eulerr package) (2019-05-09, Thu)
   - <https://github.com/GuangchuangYu/ggplotify/issues/5>
 
 # ggplotify 0.0.3